Deprecate the cURL checks in Site Health
